requirements.txt
google-adk>=0.2.0
google-genai>=1.5.0
selenium>=4.30.0
webdriver-manager>=4.0.2
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
tenacity>=8.2.0
python-dateutil>=2.8.2
lxml>=4.9.0
fastapi>=0.115.0
pydantic>=2.0.0
uvicorn>=0.34.0
coloredlogs>=15.0.1
concurrent-log-handler>=0.9.24
streamlit>=1.31.0
streamlit-autorefresh>=1.0.0
networkx>=3.2.1
matplotlib>=3.8.0